backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
python-dotenv
requests
joblib
APScheduler
xgboost==3.0.5
pytest==8.4.2
httpx==0.28.1